client: prevent leak of file descriptors with multiple clients
Each client keeps alive up to 2 connections, so the number of clients created in
the lifetime of an application must be constant to prevent exhaustion of file
descriptors.  One way to fix this is keep a hidden constant pool of underlying
http clients, and another is not to keep connections open.